Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Cifrario: Sistema para la gestión de horas sociales para instituciones"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.1",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "mysqlclient>=2.2.7",
    "pillow>=11.3.0",
    "whitenoise>=6.9.0",
]
